+++ test/Sentry.Extensions.Logging.Tests/SentryLoggingOptionsSetupTests.cs
@@ -57,6 +57,8 @@ public class SentryLoggingOptionsSetupTests
             MinimumEventLevel = LogLevel.Error,
             InitializeSdk = true
         };
+        expected.Experimental.EnableLogs = true;
+        expected.ExperimentalLogging.MinimumLogLevel = LogLevel.None;
         var config = new ConfigurationBuilder()
             .AddInMemoryCollection(new Dictionary<string, string>
             {
@@ -106,6 +108,9 @@ public class SentryLoggingOptionsSetupTests
                 ["MinimumBreadcrumbLevel"] = expected.MinimumBreadcrumbLevel.ToString(),
                 ["MinimumEventLevel"] = expected.MinimumEventLevel.ToString(),
                 ["InitializeSdk"] = expected.InitializeSdk.ToString(),
+
+                ["Experimental:EnableLogs"] = expected.Experimental.EnableLogs.ToString(),
+                ["ExperimentalLogging:MinimumLogLevel"] = expected.ExperimentalLogging.MinimumLogLevel.ToString(),
             })
             .Build();
 
@@ -163,6 +168,9 @@ public class SentryLoggingOptionsSetupTests
             actual.MinimumBreadcrumbLevel.Should().Be(expected.MinimumBreadcrumbLevel);
             actual.MinimumEventLevel.Should().Be(expected.MinimumEventLevel);
             actual.InitializeSdk.Should().Be(expected.InitializeSdk);
+
+            actual.Experimental.EnableLogs.Should().Be(expected.Experimental.EnableLogs);
+            actual.ExperimentalLogging.MinimumLogLevel.Should().Be(expected.ExperimentalLogging.MinimumLogLevel);
         }
     }
 }