+++ test/Sentry.Extensions.Logging.Tests/SentryStructuredLoggerTests.cs
@@ -1,0 +1,38 @@
+#nullable enable
+
+using Microsoft.Extensions.Logging;
+
+namespace Sentry.Extensions.Logging.Tests;
+
+public class SentryStructuredLoggerTests
+{
+    [Fact]
+    public void SmokeTest()
+    {
+        InMemorySentryStructuredLogger inMemory = new();
+
+        string categoryName = "CategoryName";
+        SentryLoggingOptions options = new();
+        options.Experimental.EnableLogs = true;
+        IHub hub = Substitute.For<IHub>();
+        hub.IsEnabled.Returns(true);
+        hub.Logger.Returns(inMemory);
+
+        var logger = new SentryStructuredLogger(categoryName, options, hub);
+
+        IDisposable? disposable = logger.BeginScope("state");
+        disposable.Should().NotBeNull();
+
+        logger.IsEnabled(LogLevel.Warning).Should().BeTrue();
+
+        EventId eventId = new(1, "eventId");
+        Exception exception = new InvalidOperationException();
+        Func<string, Exception?, string> formatter = (string state, Exception? exception) =>
+        {
+            state.Should().Be("state");
+            exception.Should().BeOfType<InvalidOperationException>();
+            return "Message";
+        };
+        logger.Log(LogLevel.Warning, eventId, "state", exception, formatter);
+    }
+}